format

+++ test.ts
@@ -1,8 +1,8 @@
-import Path from "path";
-import Fs from "fs";
-import Colors from "colors/safe";
+import { Database, Table, Util } from "./index";
 
-import { Util, Database, Table } from "./index";
+import Colors from "colors/safe";
+import Fs from "fs";
+import Path from "path";
 
 function log(message: string, ...parts: any[]) {
     logColored("gray", message, ...parts);
@@ -15,8 +15,7 @@ function logStep(stepName: string) {
     log(`
 ###
 # Current step: ${stepName}
-###`
-    )
+###`);
 }
 
 function ok() {
@@ -26,8 +25,7 @@ function ok() {
 function assert(a: any, b: any) {
     log("comparing plain values:", a, b);
 
-    if (a != b)
-	throw `${a} does not match ${b}`;
+    if (a != b) throw `${a} does not match ${b}`;
 
     ok();
 }
@@ -35,17 +33,15 @@ function assert(a: any, b: any) {
 function assertArrays(a: any[], b: any[]) {
     log("comparing arrays:", a, b);
 
-    if (a.length != b.length) 
-	throw "array sizes do not mach"
+    if (a.length != b.length) throw "array sizes do not mach";
 
     for (let i = 0; i < a.length; i++) {
-	const aItem = a[i];
-	const bItem = b[i];
+        const aItem = a[i];
+        const bItem = b[i];
 
-	if (aItem == bItem) 
-	    continue;
+        if (aItem == bItem) continue;
 
-	throw `${aItem} does not match ${bItem} (index ${i}`;
+        throw `${aItem} does not match ${bItem} (index ${i}`;
     }
 
     ok();
@@ -56,9 +52,8 @@ const TEST_DIR = "test";
 const DB_BASE = Path.join(TEST_DIR, "base");
 
 log("preparing...");
-if (Fs.existsSync(TEST_DIR))
-    Fs.rmSync(TEST_DIR, { recursive: true });
-Fs.mkdirSync(TEST_DIR, { recursive: true })
+if (Fs.existsSync(TEST_DIR)) Fs.rmSync(TEST_DIR, { recursive: true });
+Fs.mkdirSync(TEST_DIR, { recursive: true });
 
 log("starting test...");
 
@@ -146,28 +141,52 @@ logStep("table basic");
     assertArrays(table.entryContainerPath, [TABLE_NAME, "entries"]);
 
     log("checking specific field path...");
-    // @ts-expect-error 
-    assertArrays(table.getPathForField(FIELD_NAME), [TABLE_NAME, "fields", FIELD_NAME]);
+    // @ts-expect-error
+    assertArrays(table.getPathForField(FIELD_NAME), [
+        TABLE_NAME,
+        "fields",
+        FIELD_NAME,
+    ]);
 
     log("checking specific field value path...");
-    // @ts-expect-error 
-    assertArrays(table.getValuePathForField(FIELD_NAME, VALUE_NAME), [TABLE_NAME, "fields", FIELD_NAME, VALUE_KEY]);
+    // @ts-expect-error
+    assertArrays(table.getValuePathForField(FIELD_NAME, VALUE_NAME), [
+        TABLE_NAME,
+        "fields",
+        FIELD_NAME,
+        VALUE_KEY,
+    ]);
 
     log("checking specific entry path in field...");
-    // @ts-expect-error 
-    assertArrays(table.getEntryPathForFieldValue(FIELD_NAME, VALUE_NAME, ENTRY_ID), [TABLE_NAME, "fields", FIELD_NAME, VALUE_KEY, ENTRY_ID]);
+    // @ts-expect-error
+    assertArrays(
+        table.getEntryPathForFieldValue(FIELD_NAME, VALUE_NAME, ENTRY_ID),
+        [TABLE_NAME, "fields", FIELD_NAME, VALUE_KEY, ENTRY_ID],
+    );
 
     log("checking specific entry path...");
-    // @ts-expect-error 
-    assertArrays(table.getPathForEntry(ENTRY_ID), [TABLE_NAME, "entries", ENTRY_ID]);
+    // @ts-expect-error
+    assertArrays(table.getPathForEntry(ENTRY_ID), [
+        TABLE_NAME,
+        "entries",
+        ENTRY_ID,
+    ]);
 
     log("checking specific field path in entry...");
-    // @ts-expect-error 
-    assertArrays(table.getFieldPathForEntry(ENTRY_ID, FIELD_NAME), [TABLE_NAME, "entries", ENTRY_ID, FIELD_NAME]);
+    // @ts-expect-error
+    assertArrays(table.getFieldPathForEntry(ENTRY_ID, FIELD_NAME), [
+        TABLE_NAME,
+        "entries",
+        ENTRY_ID,
+        FIELD_NAME,
+    ]);
 
     log("checking specific field value path in entry...");
-    // @ts-expect-error 
-    assertArrays(table.getFieldValuePathForEntry(ENTRY_ID, FIELD_NAME, VALUE_NAME), [TABLE_NAME, "entries", ENTRY_ID, FIELD_NAME, VALUE_KEY]);
+    // @ts-expect-error
+    assertArrays(
+        table.getFieldValuePathForEntry(ENTRY_ID, FIELD_NAME, VALUE_NAME),
+        [TABLE_NAME, "entries", ENTRY_ID, FIELD_NAME, VALUE_KEY],
+    );
 }
 
 /////
@@ -177,72 +196,81 @@ logStep("core");
     const COMPLEX_VALUE = "value,./;:'[]{}()!@#$%^&*";
 
     async function testTable(tableName: string) {
-	const table = new Table(tableName, database);
+        const table = new Table(tableName, database);
 
-	async function testEntry(suffix: string) {
-	    const entryId = `entry-${suffix}`;
+        async function testEntry(suffix: string) {
+            const entryId = `entry-${suffix}`;
 
-	    log("adding values to fields...");
-	    await table.addFieldValuesToEntry(entryId, "a", ["a", "b", "c"]);
-	    await table.addFieldValuesToEntry(entryId, "a", ["c", "d"]);
-	    await table.addFieldValuesToEntry(entryId, "b", ["1", "2", COMPLEX_VALUE]);
-	    await table.addFieldValuesToEntry(entryId, "id", [suffix]);
+            log("adding values to fields...");
+            await table.addFieldValuesToEntry(entryId, "a", ["a", "b", "c"]);
+            await table.addFieldValuesToEntry(entryId, "a", ["c", "d"]);
+            await table.addFieldValuesToEntry(entryId, "b", [
+                "1",
+                "2",
+                COMPLEX_VALUE,
+            ]);
+            await table.addFieldValuesToEntry(entryId, "id", [suffix]);
 
-	    log("getting fields...");
-	    const fields = await table.getFieldsOfEntry(entryId);
-	    assertArrays(fields, ["a", "b", "id"]);
+            log("getting fields...");
+            const fields = await table.getFieldsOfEntry(entryId);
+            assertArrays(fields, ["a", "b", "id"]);
 
-	    log("getting values...");
-	    const aValues = await table.getValuesForField(entryId, "a");
-	    const bValues = await table.getValuesForField(entryId, "b");
-	    const unknownValues = await table.getValuesForField(entryId, "x");
-	    assertArrays(aValues, ["a", "b", "c", "d"]);
-	    assertArrays(bValues, ["1", "2", COMPLEX_VALUE]);
-	    assertArrays(unknownValues, []);
+            log("getting values...");
+            const aValues = await table.getValuesForField(entryId, "a");
+            const bValues = await table.getValuesForField(entryId, "b");
+            const unknownValues = await table.getValuesForField(entryId, "x");
+            assertArrays(aValues, ["a", "b", "c", "d"]);
+            assertArrays(bValues, ["1", "2", COMPLEX_VALUE]);
+            assertArrays(unknownValues, []);
 
-	    log("deleting values...");
-	    await table.removeFieldValuesFromEntry(entryId, "a", ["a", "b"]);
-	    const removeControl = await table.getValuesForField(entryId, "a");
-	    assertArrays(removeControl, ["c", "d"]);
+            log("deleting values...");
+            await table.removeFieldValuesFromEntry(entryId, "a", ["a", "b"]);
+            const removeControl = await table.getValuesForField(entryId, "a");
+            assertArrays(removeControl, ["c", "d"]);
 
-	    log("clearing values...");
-	    await table.clearFieldValuesForEntry(entryId, "b");
-	    const clearControl = await table.getValuesForField(entryId, "b");
-	    assertArrays(clearControl, []);
+            log("clearing values...");
+            await table.clearFieldValuesForEntry(entryId, "b");
+            const clearControl = await table.getValuesForField(entryId, "b");
+            assertArrays(clearControl, []);
 
-	    log("setting values...");
-	    await table.setFieldValuesForEntry(entryId, "a", ["1", "2"]);
-	    const setControl = await table.getValuesForField(entryId, "a");
-	    assertArrays(setControl, ["1", "2"]);
+            log("setting values...");
+            await table.setFieldValuesForEntry(entryId, "a", ["1", "2"]);
+            const setControl = await table.getValuesForField(entryId, "a");
+            assertArrays(setControl, ["1", "2"]);
 
-	    log("removing non-existent entry...");
-	    await table.removeEntry("x");
+            log("removing non-existent entry...");
+            await table.removeEntry("x");
 
-	    log("accessing non-existent field...");
-	    await table.getValuesForField("x", "x");
+            log("accessing non-existent field...");
+            await table.getValuesForField("x", "x");
 
-	    log("deleting non-existent value...");
-	    await table.removeFieldValuesFromEntry("x", "x", ["x"]);
-	}
+            log("deleting non-existent value...");
+            await table.removeFieldValuesFromEntry("x", "x", ["x"]);
+        }
 
-	await testEntry("1");
-	await testEntry("2");
+        await testEntry("1");
+        await testEntry("2");
 
-	log("getting all entries...");
-	const entries = await table.getAllEntries();
-	assertArrays(entries, ["entry-1", "entry-2"]);
-	
-	log("getting entries by field a=1...");
-	const matchingEntries1 = await table.getEntriesByFieldValue("a", ["1"])
-	assertArrays(matchingEntries1, ["entry-1", "entry-2"]);
-	
-	log("getting entries by field id=1|2...");
-	const matchingEntries2 = await table.getEntriesByFieldValue("id", ["1", "2"])
-	assertArrays(matchingEntries2, ["entry-1", "entry-2"]);
-	
-	log("getting entries by field id=1...");
-	const matchingEntries3 = await table.getEntriesByFieldValue("id", ["1"])
-	assertArrays(matchingEntries3, ["entry-1"]);
+        log("getting all entries...");
+        const entries = await table.getAllEntries();
+        assertArrays(entries, ["entry-1", "entry-2"]);
+
+        log("getting entries by field a=1...");
+        const matchingEntries1 = await table.getEntriesByFieldValue("a", ["1"]);
+        assertArrays(matchingEntries1, ["entry-1", "entry-2"]);
+
+        log("getting entries by field id=1|2...");
+        const matchingEntries2 = await table.getEntriesByFieldValue("id", [
+            "1",
+            "2",
+        ]);
+        assertArrays(matchingEntries2, ["entry-1", "entry-2"]);
+
+        log("getting entries by field id=1...");
+        const matchingEntries3 = await table.getEntriesByFieldValue("id", [
+            "1",
+        ]);
+        assertArrays(matchingEntries3, ["entry-1"]);
     }
 
     await testTable("A");
@@ -251,8 +279,8 @@ logStep("core");
 
 function testTypes() {
     enum FieldNames {
-	"a", 
-	"b"
+        "a",
+        "b",
     }
 
     const database = new Database(".");
